Change touch-action CSS to manipulation instead of none
touch-action: none prevents scrolling in the puzzle list view. Supposedly
touch-action: manipulation should only disable double-tap zoom. But either way
it's kind of pointless because fucking iOS doesn't disable double-tap zoom-in.
Zoom-out is disabled no matter the value? I have to assume this is an iOS/Safari
bug, because it just stays zoomed in and the only way to zoom out is to zoom in
and back out in the page settings. Super annoying. Maybe there's some dumb meta
tag or something I need to set, but I don't have the patience to trawl around
the internet endlessly looking for the right magic beans to plant here, when
the behavior I want seems pretty simple. Oh well, I guess the workaround is
"don't fucking double tap on iOS".